requirements.txt
django==3.0.3
djangorestframework==3.11.0
django-filter==2.2.0
django-crispy-forms==1.8.1
psycopg2==2.8.4
